% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_required_packages.R
\name{install_required_packages}
\alias{install_required_packages}
\title{Install Required Packages}
\usage{
install_required_packages(package_name, source)
}
\arguments{
\item{package_name}{The name of the package to check/install.}

\item{source}{The source of the package (CRAN, github::, or url:: for specific version).}
}
\description{
This function checks if a specified package is installed and its version, and installs it if necessary.
}
\examples{
install_required_packages("tidyverse", "CRAN")
install_required_packages("hud.keep", "github::hud-govt-nz/hud-keep")
install_required_packages("dplyr", "url::https://cran.r-project.org/src/contrib/Archive/dplyr/dplyr_1.1.3.tar.gz")
}
