% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{libraries}
\alias{libraries}
\title{Libraries}
\usage{
libraries(packages)
}
\arguments{
\item{packages}{A named list where names are package names and values are their sources (CRAN, github::, or url::).}
}
\description{
This function will install (if required), then load a list of packages and validates their versions if specified.
}
