% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_required_packages.R
\name{load_required_packages}
\alias{load_required_packages}
\title{Load Required Packages}
\usage{
load_required_packages(packages)
}
\arguments{
\item{packages}{A named list where names are package names and values are their sources (CRAN, github::, or url::).}
}
\description{
This function loads a list of packages and checks their versions if specified.
}
